moved helper functions to ServerUtils.cpp file
