+++ .gitignore
@@ -7,6 +7,9 @@ issues
 # Ignore settings
 .vscode/settings.json
 
+# ignore debug
+debugger.py
+
 # Created by https://www.toptal.com/developers/gitignore/api/visualstudiocode,clion,sublimetext,c++,vim
 # Edit at https://www.toptal.com/developers/gitignore?templates=visualstudiocode,clion,sublimetext,c++,vim
