--- include/ParsingUtils.hpp
@@ -1,8 +1,0 @@
-#ifndef PARSING_UTILS_HPP
-#define PARSING_UTILS_HPP
-
-#include <string>
-
-std::string trimSpace(std::string s);
-
-#endif
