+++ server/server.cfg
@@ -14,7 +14,7 @@ mime {
 # ALPHA server
 server {
 	host localhost
-	port 3434
+	port 2425
 	server_names ["alpha"]
 
 	error_page 403 "error_pages/403.html"
@@ -61,7 +61,7 @@ server {
 # BETA server
 server {
 	host localhost
-	port 3434
+	port 2425
 	server_names ["beta"]
 
 	error_page 403 "error_pages/403.html"
@@ -88,7 +88,7 @@ server {
 # GAMMA server
 server {
 	host localhost
-	port 6623
+	port 5432
 	server_names ["alpha"]
 
 	error_page 403 "error_pages/403.html"
