+++ src/Server.cpp
@@ -2,6 +2,7 @@
 #include "ServerConfig.hpp"
 #include "HttpRequest.hpp"
 #include "HttpResponse.hpp"
+#include "ServerUtils.hpp"
 #include <iostream>
 #include <unistd.h>
 #include <fcntl.h>
@@ -10,16 +11,9 @@
 #include <algorithm>
 #include <string>
 
-// POSIX for @is_file, @is_directory and @list_dir
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <dirent.h>
+Server::Server( ServerCfg& cfg, ServerConfig& gen_cfg ) :_cfg( cfg ), _gen_cfg( gen_cfg ) {}
 
-Server::Server( ServerCfg& cfg, ServerConfig& gen_cfg ) :_cfg( cfg ), _gen_cfg( gen_cfg ) {
-}
-
-Server::Server( const Server& copy ) :_cfg( copy._cfg ), _gen_cfg( copy._gen_cfg ) {
-}
+Server::Server( const Server& copy ) :_cfg( copy._cfg ), _gen_cfg( copy._gen_cfg ) {}
 
 Server::~Server() { }
 
@@ -45,122 +39,6 @@ RouteCfg* find_route(const HttpRequest& req, std::vector<RouteCfg>& routes)
     return route_match.second;
 }
 
-/*
-*   @is_accepted_method:
-*    Checks if the requested method is accepted by the route
-*/
-bool is_accepted_method(RouteCfg* route, const std::string method) {
-    return std::find(route->accepted_methods.begin(), route->accepted_methods.end(), method) != route->accepted_methods.end();
-}
-
-
-/*
-*   @is_file:
-*       Checks if given path points to a file
-*       Path must not end in '/'
-*/
-bool    is_file( const std::string& path ) {
-    // Path must not end in '/'
-    if ( path[path.size() - 1] == '/' ) return false;
-
-    struct stat buf;
-    
-    // Any error returns false
-    if ( stat(path.c_str(), &buf) != 0 ) return false;
-
-    // Check if is a file
-    return S_ISREG(buf.st_mode);
-}
-
-/*
-*   @is_directory:
-*/
-bool is_directory(const std::string& path) {
-    struct stat buf;
-    
-    // Any error returns false
-    if ( stat(path.c_str(), &buf) != 0 ) return false;
-    
-    // Check if is a dir
-    return S_ISDIR(buf.st_mode);
-}
-
-/*
-*   @list_dir:
-*       Returns a vector of strings with all files/folders inside a dir
-*/
-std::vector<std::string>    list_dir( const std::string& path ) {
-    std::vector<std::string>    dir_listing;
-
-    DIR* dir;
-    struct dirent* ent;
-
-    dir = opendir( path.c_str() );
-    
-    if ( dir != NULL ) {
-        while ( ( ent = readdir( dir ) ) != NULL ) {
-            dir_listing.push_back( ent->d_name );
-        }
-        closedir(dir);
-    }
-    return dir_listing;
-}
-
-#include <iostream>
-/*
-*   @get_path:
-*       If auto_index is true, returns user request
-*       If not, return predefined index
-*
-*/
-int get_path(const HttpRequest& req, RouteCfg* route, std::string& path ) {
-    // If root ends in '/', remove last char
-    if (!route->root.empty() && route->root[route->root.size() - 1] == '/')
-        route->root = route->root.substr(0, route->root.size() - 1);
-
-    // Request is equal to relative path ('.') + root path + route path
-    path = route->root + req.target();
-
-    // if is a file return request
-    if ( is_file( path ) ) {
-        return 0;
-    }
-    else if ( is_directory( path ) ) {
-        std::vector<std::string>    dir_listing = list_dir( path );
-        
-        // If index.html exists in said folder, return request
-        std::vector<std::string>::iterator it = std::find(dir_listing.begin(), dir_listing.end(), "index.html");
-        if ( it != dir_listing.end() ) {
-            if (!route->route_path.empty() && route->route_path[route->route_path.size() - 1] == '/')
-                route->route_path = route->route_path.substr(0, route->route_path.size() - 1);
-            path = route->root + route->route_path + "/" + "index.html";
-            return 0;
-        } else if ( route->auto_index ) { // else if (auto_index on), return list of contents
-            return 1;
-        }
-    }
-
-    path = "";
-    return 2;
-}
-
-/*
-*   @replace_occurrence:
-*       Replace all uccurrences of str2 by str3 in str1
-*/
-int replace_occurrence( std::string& str, const std::string& occurr, const std::string& replacement) {
-    size_t start_pos = 0;
-    int i = 0;
-
-    while( ( start_pos = str.find( occurr, start_pos)) != std::string::npos ) {
-        str.replace( start_pos, occurr.length(), replacement );
-        start_pos += replacement.length();
-        i++;
-    }
-    
-    return i;
-}
-
 HttpResponse    list_dir_res( HttpResponse& res, const HttpRequest& req, std::string path) {
         res.body().clear();
         
@@ -175,7 +53,7 @@ HttpResponse    list_dir_res( HttpResponse& res, const HttpRequest& req, std::st
         
         std::string                 items;
         std::vector<std::string>    dir_listing = list_dir( path );
-        for (size_t i = 0; i < dir_listing.size(); i++)
+        for ( size_t i = 0; i < dir_listing.size(); i++ )
             items.append( "<li><a href=\"" + req.target() + "/" + dir_listing[i] + "\">" + dir_listing[i] + "</a></li>" );
 
         // Replace all occurs of [DIR] & [ITEMS] with dir name and dir listing
@@ -185,8 +63,8 @@ HttpResponse    list_dir_res( HttpResponse& res, const HttpRequest& req, std::st
             res.body().append( line_buff );
         }
 
-        res.set_status(200, "OK");
-        res.set_header("Content-Type", "text/html");
+        res.set_status( 200, "OK" );
+        res.set_header( "Content-Type", "text/html" );
         file.close();
         return res;
 }
@@ -251,7 +129,7 @@ HttpResponse Server::handleRequest(const HttpRequest& req)
         ss << res.body().length();
         res.set_status(200, "OK");
         //TODO add MIME type
-        res.set_header("Content-Type", "text/plain");
+        res.set_header("Content-Type", "text/html");
     }
     return res;
 }
